Finished segmentation of examples
All examples are now in their own file, in their own function
