+++ examples/argsExamples/example12.py
@@ -1,20 +1,20 @@
 from raytracing import ImagingPath, Space, Lens
 from matplotlib import pyplot as plt
 
-'''An object at z=0 (front edge) is used. It is shown in blue. The image (or any intermediate images) are shown in red.\n\
-This will use the default objectHeight and fanAngle but they can be changed with:
-path.objectHeight = 1.0
-path.fanAngle = 0.5
-path.fanNumber = 5
-path.rayNumber = 3'''
+'''
+Demo #12: Thick diverging lens built from individual elements
+'''
 
 
 def example():
     path = ImagingPath()
-    path.label = "Demo #1: lens f = 5cm, infinite diameter"
-    path.append(Space(d=10))
-    path.append(Lens(f=5))
-    path.append(Space(d=10))
+    path.label = "Demo #12: Thick diverging lens built from individual elements"
+    path.objectHeight = 20
+    path.append(Space(d=50))
+    path.append(DielectricInterface(R=-20, n1=1.0, n2=1.55, diameter=25, label='Front'))
+    path.append(Space(d=10, diameter=25, label='Lens'))
+    path.append(DielectricInterface(R=20, n1=1.55, n2=1.0, diameter=25, label='Back'))
+    path.append(Space(d=50))
     fig, axes = plt.subplots(figsize=(10, 7))
     path.createRayTracePlot(axes=axes)
     plt.savefig('tempFig.pdf', dpi=600)
