+++ examples/argsExamples/example19.py
@@ -1,26 +1,33 @@
 from raytracing import ImagingPath, Space, Lens
 from matplotlib import pyplot as plt
 
-'''An object at z=0 (front edge) is used. It is shown in blue. The image (or any intermediate images) are shown in red.\n\
-This will use the default objectHeight and fanAngle but they can be changed with:
-path.objectHeight = 1.0
-path.fanAngle = 0.5
-path.fanNumber = 5
-path.rayNumber = 3'''
+'''Demo #19 - Calculation of cavity laser modes'''
 
 
 def example():
-    path = ImagingPath()
-    path.label = "Demo #1: lens f = 5cm, infinite diameter"
-    path.append(Space(d=10))
-    path.append(Lens(f=5))
-    path.append(Space(d=10))
+    cavity = LaserPath(label="Laser cavity: round trip\nCalculated laser modes")
+    cavity.isResonator = True
+    cavity.append(Space(d=160))
+    cavity.append(DielectricSlab(thickness=100, n=1.8))
+    cavity.append(Space(d=160))
+    cavity.append(CurvedMirror(R=400))
+    cavity.append(Space(d=160))
+    cavity.append(DielectricSlab(thickness=100, n=1.8))
+    cavity.append(Space(d=160))
     fig, axes = plt.subplots(figsize=(10, 7))
-    path.createRayTracePlot(axes=axes)
+    cavity.createRayTracePlot(axes=axes)
     plt.savefig('tempFig.pdf', dpi=600)
 
+    # Calculate all self-replicating modes (i.e. eigenmodes)
+    (q1, q2) = cavity.eigenModes()
+    print(q1, q2)
+    # Obtain all physical (i.e. finite) self-replicating modes
+    qs = cavity.laserModes()
+    for q in qs:
+        print(q)
+
     if __name__ == "__main__":
-        path._showPlot()
+        cavity._showPlot()
 
 
 if __name__ == "__main__":
