+++ examples/argsExamples/example6.py
@@ -1,19 +1,18 @@
 from raytracing import ImagingPath, Space, Lens, Aperture
 from matplotlib import pyplot as plt
 
-'''# Demo #5: Simple microscope system
-The aperture stop (AS) is at the entrance of the objective lens, and the tube lens, in this particular microscope, is
-the field stop (FS) and limits the field of view. Because the field stop exists, we can use limitObjectToFieldOfView=True
-when displaying, which will set the objectHeight to the field of view, but will still trace all the rays using our parameters.
+'''
+    # Demo #6: Simple microscope system, only principal rays
+    The aperture stop (AS) is at the entrance of the objective lens, and the tube lens, in this particular microscope, is
+    the field stop (FS) and limits the field of view. Because the field stop exists, we can use limitObjectToFieldOfView=True
+    when displaying, which will set the objectHeight to the field of view. We can also require that only the principal rays are drawn: chief ray
+    marginal ray (or axial ray).
 '''
 
 
 def example():
     path = ImagingPath()
-    path.label = "Demo #5: Simple microscope system"
-    path.fanAngle = 0.1  # full fan angle for rays
-    path.fanNumber = 5  # number of rays in fan
-    path.rayNumber = 5  # number of points on object
+    path.label = "Demo #6: Simple microscope system, only principal rays"
     path.append(Space(d=4))
     path.append(Lens(f=4, diameter=0.8, label='Obj'))
     path.append(Space(d=4 + 18))
